Page: Remove _app and _document and add feature to disable scroll
