--- src/pages/_document.js
@@ -1,30 +1,0 @@
-/* eslint-disable react/no-danger */
-import React from 'react';
-import Document, { Head, Main, NextScript } from 'next/document';
-
-export default class MyDocument extends Document {
-  render() {
-    return (
-      <html lang="en">
-        <Head>
-          <meta charSet="utf-8" />
-          <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <link rel="shortcut icon" href="/static/favicon.ico" />
-          <link
-            rel="stylesheet"
-            href="https://unpkg.com/tachyons@4.11.1/css/tachyons.min.css"
-          />
-          <link
-            rel="stylesheet"
-            href="https://fonts.googleapis.com/css?family=Lobster"
-          />
-        </Head>
-        <body className="sans-serif" style={{ fontSize: '18px' }}>
-          <Main />
-          <NextScript />
-        </body>
-      </html>
-    );
-  }
-}
